+++ maps-kt-geojson/src/commonMain/kotlin/space/kscience/maps/geojson/GeoJsonPropertiesAttribute.kt
@@ -2,6 +2,6 @@ package space.kscience.maps.geojson
 
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.serializer
-import space.kscience.SerializableAttribute
+import space.kscience.attributes.SerializableAttribute
 
 public object GeoJsonPropertiesAttribute : SerializableAttribute<JsonObject>("properties", serializer())