+++ trajectory-kt/src/commonTest/kotlin/space/kscience/kmath/geometry/CircleTests.kt
@@ -5,14 +5,14 @@
 
 package space.kscience.kmath.geometry
 
-import space.kscience.circle
-import space.kscience.intersects
-import space.kscience.intersectsOrInside
 import space.kscience.kmath.geometry.euclidean2d.Circle2D
 import space.kscience.kmath.geometry.euclidean2d.Float64Space2D
 import space.kscience.kmath.geometry.euclidean2d.circumference
 import space.kscience.kmath.structures.Float64
-import space.kscience.segment
+import space.kscience.trajectory.circle
+import space.kscience.trajectory.intersects
+import space.kscience.trajectory.intersectsOrInside
+import space.kscience.trajectory.segment
 import kotlin.math.pow
 import kotlin.math.sqrt
 import kotlin.test.Test